adopt shared e2e helpers from serverless-ci
